fix(admin): inline preview controls, drop the page-level toolbar
User feedback: the standalone "Live preview" toolbar above the editor
sections felt disconnected from the panes it was driving. The page-level
banner is gone; each preview pane now carries the controls in its own
header — operator-grade UI, no global widget.

Per pane header:
  Left:  "LIVE PREVIEW" caption + status badge (Live/Syncing/Stale/Error)
  Right: Sample-variables button (opens a popover with the var inputs)
         + Refresh button (bypasses the debounce)

Sample-variables popover shows the same map across every pane (the
single source of truth in the parent state); editing in any pane's
popover updates all three previews after the same debounce window.
No "Apply" button — the debounce is the apply.

Removed: MailTemplatePreviewToolbar component + 4 specs.
Adapted: MailTemplatePreviewPane gains 4 props (sampleVars,
onSampleVarsChange, onRefresh, placeholders) and renders the controls
inline. EmailTemplateEditPage drops the toolbar render and forwards the
shared state to each pane.

Tests: 379 green (was 381 — net -2 toolbar specs + 2 new pane-header
specs).
